+++ firebase-auth/src/commonMain/kotlin/dev/gitlive/firebase/auth/multifactor.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2020 GitLive Ltd.  Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package dev.gitlive.firebase.auth
+
+expect class MultiFactor {
+    val enrolledFactors: List<MultiFactorInfo>
+    suspend fun enroll(multiFactorAssertion: MultiFactorAssertion, displayName: String?)
+    suspend fun getSession(): MultiFactorSession
+    suspend fun unenroll(multiFactorInfo: MultiFactorInfo)
+    suspend fun unenroll(factorUid: String)
+}
+
+expect class MultiFactorInfo {
+    val displayName: String?
+    val enrollmentTime: Long
+    val factorId: String
+    val uid: String
+}
+
+expect class MultiFactorAssertion {
+    val factorId: String
+}
+
+expect class MultiFactorSession
