+++ firebase-auth/src/iosMain/kotlin/dev/gitlive/firebase/auth/auth.kt
@@ -103,155 +103,6 @@ actual class ActionCodeResult(val ios: FIRActionCodeInfo) {
 
 actual class SignInMethodQueryResult(actual val signInMethods: List<String>)
 
-actual class FirebaseUser internal constructor(val ios: FIRUser) {
-    actual val uid: String
-        get() = ios.uid
-    actual val displayName: String?
-        get() = ios.displayName
-    actual val email: String?
-        get() = ios.email
-    actual val phoneNumber: String?
-        get() = ios.phoneNumber
-    actual val photoURL: String?
-        get() = ios.photoURL?.absoluteString
-    actual val isAnonymous: Boolean
-        get() = ios.anonymous
-    actual val isEmailVerified: Boolean
-        get() = ios.emailVerified
-    actual val metaData: MetaData?
-        get() = MetaData(ios.metadata)
-    actual val multiFactor: MultiFactor
-        get() = MultiFactor(ios.multiFactor)
-    actual val providerData: List<UserInfo>
-        get() = ios.providerData.mapNotNull { provider -> (provider as? FIRUserInfoProtocol)?.let { UserInfo(it) } }
-    actual val providerId: String
-        get() = ios.providerID
-    actual suspend fun delete() = ios.await { deleteWithCompletion(it) }.run { Unit }
-    actual suspend fun reload() = ios.await { reloadWithCompletion(it) }.run { Unit }
-    actual suspend fun getIdToken(forceRefresh: Boolean) {
-        val token: String = ios.awaitResult { getIDTokenForcingRefresh(forceRefresh, it) }
-    }
-    actual suspend fun linkWithCredential(credential: AuthCredential): AuthResult = AuthResult(ios.awaitResult { linkWithCredential(credential.ios, it) })
-    actual suspend fun reauthenticate(credential: AuthCredential) {
-        val result: FIRAuthDataResult = ios.awaitResult { reauthenticateWithCredential(credential.ios, it) }
-    }
-    actual suspend fun reauthenticateAndRetrieveData(credential: AuthCredential): AuthResult = AuthResult(ios.awaitResult { reauthenticateAndRetrieveDataWithCredential(credential.ios, it) })
-
-    actual suspend fun sendEmailVerification(actionCodeSettings: ActionCodeSettings?) = ios.await {
-        actionCodeSettings?.let { actionSettings -> sendEmailVerificationWithActionCodeSettings(actionSettings.ios, it) } ?: sendEmailVerificationWithCompletion(it)
-    }.run { Unit }
-    actual suspend fun unlink(provider: String): FirebaseUser? {
-        val user: FIRUser? = ios.awaitResult { unlinkFromProvider(provider, it) }
-        return user?.let {
-            FirebaseUser(it)
-        }
-    }
-    actual suspend fun updateEmail(email: String) = ios.await { updateEmail(email, it) }.run { Unit }
-    actual suspend fun updatePassword(password: String) = ios.await { updatePassword(password, it) }.run { Unit }
-    actual suspend fun updatePhoneNumber(credential: PhoneAuthCredential) = ios.await { updatePhoneNumberCredential(credential.ios, it) }.run { Unit }
-    actual suspend fun updateProfile(buildRequest: (UserProfileChangeRequest.Builder) -> Unit) {
-        val request = UserProfileChangeRequest.Builder(this.ios).apply { buildRequest(this) }.build()
-        ios.await { request.ios.commitChangesWithCompletion(it) }
-    }
-    actual suspend fun verifyBeforeUpdateEmail(newEmail: String, actionCodeSettings: ActionCodeSettings?) = ios.await {
-        actionCodeSettings?.let { actionSettings -> sendEmailVerificationBeforeUpdatingEmail(newEmail, actionSettings.ios, it) } ?: sendEmailVerificationBeforeUpdatingEmail(newEmail, it)
-    }.run { Unit }
-}
-
-actual class UserInfo(val ios: FIRUserInfoProtocol) {
-    actual val displayName: String?
-        get() = ios.displayName
-    actual val email: String?
-        get() = ios.email
-    actual val phoneNumber: String?
-        get() = ios.phoneNumber
-    actual val photoURL: String?
-        get() = ios.photoURL?.absoluteString
-    actual val providerId: String
-        get() = ios.providerID
-    actual val uid: String
-        get() = ios.uid
-}
-
-actual class MetaData(val ios: FIRUserMetadata) {
-    actual val creationTime: Long?
-        get() = ios.creationDate?.timeIntervalSinceReferenceDate?.toLong()
-    actual val lastSignInTime: Long?
-        get() = ios.lastSignInDate?.timeIntervalSinceReferenceDate?.toLong()
-}
-
-actual class UserProfileChangeRequest(val ios: FIRUserProfileChangeRequest) {
-    actual class Builder(private val user: FIRUser) {
-
-        private val request = user.profileChangeRequest()
-
-        actual fun setDisplayName(displayName: String?): Builder = apply {
-            request.setDisplayName(displayName)
-        }
-        actual fun setPhotoURL(photoURL: String?): Builder = apply {
-            request.setPhotoURL(photoURL?.let { NSURL.URLWithString(it) })
-        }
-        actual fun build(): UserProfileChangeRequest = UserProfileChangeRequest(request)
-    }
-    actual val displayName: String?
-        get() = ios.displayName
-    actual val photoURL: String?
-        get() = ios.photoURL?.absoluteString
-}
-
-actual open class FirebaseAuthException(message: String): FirebaseException(message)
-actual open class FirebaseAuthActionCodeException(message: String): FirebaseAuthException(message)
-actual open class FirebaseAuthEmailException(message: String): FirebaseAuthException(message)
-actual open class FirebaseAuthInvalidCredentialsException(message: String): FirebaseAuthException(message)
-actual open class FirebaseAuthInvalidUserException(message: String): FirebaseAuthException(message)
-actual open class FirebaseAuthMultiFactorException(message: String): FirebaseAuthException(message)
-actual open class FirebaseAuthRecentLoginRequiredException(message: String): FirebaseAuthException(message)
-actual open class FirebaseAuthUserCollisionException(message: String): FirebaseAuthException(message)
-actual open class FirebaseAuthWeakPasswordException(message: String): FirebaseAuthException(message)
-actual open class FirebaseAuthWebException(message: String): FirebaseAuthException(message)
-
-actual open class AuthCredential(open val ios: FIRAuthCredential) {
-    actual val providerId: String
-        get() = ios.provider
-}
-
-actual class PhoneAuthCredential(override val ios: FIRPhoneAuthCredential) : AuthCredential(ios)
-
-actual object EmailAuthProvider {
-    actual fun credentialWithEmail(
-        email: String,
-        password: String
-    ): AuthCredential =
-        AuthCredential(FIREmailAuthProvider.credentialWithEmail(email = email, password = password))
-}
-
-actual class MultiFactor(val ios: FIRMultiFactor) {
-    actual val enrolledFactors: List<MultiFactorInfo>
-        get() = ios.enrolledFactors.mapNotNull { info -> (info as? FIRMultiFactorInfo)?.let{  MultiFactorInfo(it) } }
-    actual suspend fun enroll(multiFactorAssertion: MultiFactorAssertion, displayName: String?) = ios.await { enrollWithAssertion(multiFactorAssertion.ios, displayName, it) }.run { Unit }
-    actual suspend fun getSession(): MultiFactorSession = MultiFactorSession(ios.awaitResult { getSessionWithCompletion(completion = it) })
-    actual suspend fun unenroll(multiFactorInfo: MultiFactorInfo) = ios.await { unenrollWithInfo(multiFactorInfo.ios, it) }.run { Unit }
-    actual suspend fun unenroll(factorUid: String) = ios.await { unenrollWithFactorUID(factorUid, it) }.run { Unit }
-}
-
-actual class MultiFactorInfo(val ios: FIRMultiFactorInfo) {
-    actual val displayName: String?
-        get() = ios.displayName
-    actual val enrollmentTime: Long
-        get() = ios.enrollmentDate.timeIntervalSinceReferenceDate.toLong()
-    actual val factorId: String
-        get() = ios.factorID
-    actual val uid: String
-        get() = ios.UID
-}
-
-actual class MultiFactorAssertion(val ios: FIRMultiFactorAssertion) {
-    actual val factorId: String
-        get() = ios.factorID
-}
-
-actual class MultiFactorSession(val ios: FIRMultiFactorSession)
-
 actual class ActionCodeSettings private constructor(val ios: FIRActionCodeSettings) {
     actual class Builder(val ios: FIRActionCodeSettings = FIRActionCodeSettings()) {
         actual fun setAndroidPackageName(androidPackageName: String, installIfNotAvailable: Boolean, minimumVersion: String?): Builder = apply {
@@ -286,8 +137,18 @@ actual class ActionCodeSettings private constructor(val ios: FIRActionCodeSettin
         get() = ios.URL?.absoluteString ?: ""
 }
 
+actual open class FirebaseAuthException(message: String): FirebaseException(message)
+actual open class FirebaseAuthActionCodeException(message: String): FirebaseAuthException(message)
+actual open class FirebaseAuthEmailException(message: String): FirebaseAuthException(message)
+actual open class FirebaseAuthInvalidCredentialsException(message: String): FirebaseAuthException(message)
+actual open class FirebaseAuthInvalidUserException(message: String): FirebaseAuthException(message)
+actual open class FirebaseAuthMultiFactorException(message: String): FirebaseAuthException(message)
+actual open class FirebaseAuthRecentLoginRequiredException(message: String): FirebaseAuthException(message)
+actual open class FirebaseAuthUserCollisionException(message: String): FirebaseAuthException(message)
+actual open class FirebaseAuthWeakPasswordException(message: String): FirebaseAuthException(message)
+actual open class FirebaseAuthWebException(message: String): FirebaseAuthException(message)
 
-private fun <T, R> T.throwError(block: T.(errorPointer: CPointer<ObjCObjectVar<NSError?>>) -> R): R {
+internal fun <T, R> T.throwError(block: T.(errorPointer: CPointer<ObjCObjectVar<NSError?>>) -> R): R {
     memScoped {
         val errorPointer: CPointer<ObjCObjectVar<NSError?>> = alloc<ObjCObjectVar<NSError?>>().ptr
         val result = block(errorPointer)
@@ -299,7 +160,7 @@ private fun <T, R> T.throwError(block: T.(errorPointer: CPointer<ObjCObjectVar<N
     }
 }
 
-private suspend fun <T, R> T.awaitResult(function: T.(callback: (R?, NSError?) -> Unit) -> Unit): R {
+internal suspend fun <T, R> T.awaitResult(function: T.(callback: (R?, NSError?) -> Unit) -> Unit): R {
     val job = CompletableDeferred<R>()
     function { result, error ->
         if(result != null) {
@@ -311,7 +172,7 @@ private suspend fun <T, R> T.awaitResult(function: T.(callback: (R?, NSError?) -
     return job.await()
 }
 
-private suspend fun <T> T.await(function: T.(callback: (NSError?) -> Unit) -> Unit) {
+internal suspend fun <T> T.await(function: T.(callback: (NSError?) -> Unit) -> Unit) {
     val job = CompletableDeferred<Unit>()
     function { error ->
         if(error == null) {
@@ -342,6 +203,8 @@ private fun NSError.toException() = when(domain) {
         FIRAuthErrorCodeEmailAlreadyInUse,
         FIRAuthErrorCodeAccountExistsWithDifferentCredential,
         FIRAuthErrorCodeCredentialAlreadyInUse -> FirebaseAuthUserCollisionException(toString())
+
+        FIRAuthErrorCodeWeakPassword -> FirebaseAuthWeakPasswordException(toString())
 
         FIRAuthErrorCodeWebContextAlreadyPresented,
         FIRAuthErrorCodeWebContextCancelled,
