+++ firebase-auth/src/iosMain/kotlin/dev/gitlive/firebase/auth/credentials.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2020 GitLive Ltd.  Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package dev.gitlive.firebase.auth
+
+import cocoapods.FirebaseAuth.*
+
+actual open class AuthCredential(open val ios: FIRAuthCredential) {
+    actual val providerId: String
+        get() = ios.provider
+}
+
+actual class PhoneAuthCredential(override val ios: FIRPhoneAuthCredential) : AuthCredential(ios)
+actual class OAuthCredential(override val ios: FIROAuthCredential) : AuthCredential(ios)
+
+actual object EmailAuthProvider {
+    actual fun credentialWithEmail(
+        email: String,
+        password: String
+    ): AuthCredential =
+        AuthCredential(FIREmailAuthProvider.credentialWithEmail(email = email, password = password))
+}
+
+actual object FacebookAuthProvider {
+    actual fun credentialWithAccessToken(accessToken: String): AuthCredential = AuthCredential(FIRFacebookAuthProvider.credentialWithAccessToken(accessToken))
+}
